generic np.integer

+++ tests/test_series.py
@@ -1150,7 +1150,7 @@ def test_neg() -> None:
 def test_getattr() -> None:
     # GH 261
     series = pd.Series([1, 2, 3], index=["a", "b", "c"], dtype=int)
-    check(assert_type(series.a, int), np.intp)
+    check(assert_type(series.a, int), np.integer)
 
 
 def test_dtype_type() -> None:
